feat: Decimal item/armor weight value (#4)

+++ src/Battlescape/AIModule.cpp
@@ -6416,7 +6416,7 @@ float AIModule::getItemPickUpScore(BattleItem* item)
 	if (!valid)
 		return 0;
 	score = item->getRules()->getSellCost();
-	int mainHandWeight = 0;
+	float mainHandWeight = 0;
 	if (_unit->getMainHandWeapon())
 		mainHandWeight = _unit->getMainHandWeapon()->getRules()->getWeight();
 	float encumbrance = (float)_unit->getBaseStats()->strength / (float)(_unit->getCarriedWeight() - mainHandWeight + item->getRules()->getWeight());
